requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.17
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
phonenumbers==8.13.52
pillow==10.4.0
sqlparse==0.5.3
typing-extensions==4.12.2
